requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
SQLAlchemy==1.4.39
fastapi-crudrouter==0.8.5
alembic==1.8.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
